+++ src/ml/GenerateSchemaTypes.ml
@@ -39,10 +39,13 @@ type authorizationConfig = {
   baselinePath: string option;
 }
 
+type authorizationCoverage = AuthorizationField | AuthorizationSelection
+
 type authorizationFunctionReference = {
   path: string list;
   loc: Location.t;
   fileUri: Uri.t;
+  covers: authorizationCoverage;
 }
 
 type publicAuthorization = {reason: string; loc: Location.t; fileUri: Uri.t}
@@ -76,10 +79,16 @@ type authorizationGapKind =
 
 type effectiveAuthorizationPlan = {
   functions: authorizationFunction list;
+  inheritedSelectionPolicies: inheritedSelectionPolicy list;
   public: publicAuthorization option;
   resolverOutcome: resolverOutcome option;
   synthetic: bool;
   baselineGap: authorizationGapKind option;
+}
+
+and inheritedSelectionPolicy = {
+  boundaryCoordinate: string;
+  fn: authorizationFunction;
 }
 
 type typeLocationLoc = {